% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConnectPointsToLines.R
\name{ConnectPointsToLines}
\alias{ConnectPointsToLines}
\title{ConnectPointsToLines
For every point, a new line is created connecting it to the closest line.}
\usage{
ConnectPointsToLines(centroids, lines)
}
\arguments{
\item{centroids}{Centroids generated from the PolyCentroid function.}

\item{lines}{A spatiallines dataset or a spatiallinesdataframe.}
}
\value{
"snapped_points" A spatialpointsdataframe of the points on the line
nearest to each point.

"connector_lines" A spatiallinesdata frame. Each line connects the
snapped_points to the centroids.
}
\description{
ConnectPointsToLines
For every point, a new line is created connecting it to the closest line.
}
