% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyCentroid.R
\name{PolyCentroid}
\alias{PolyCentroid}
\title{PolyCentroid}
\usage{
PolyCentroid(polygon)
}
\arguments{
\item{A}{spatialpolygonsdataframe.}
}
\value{
A spatialpointsdataframe of the centerspoints of the polygons.
}
\description{
Take a spatialpolygonsdataframe object and generate a spatialpointsdataframe
of their spatial centers. Preserve the attributes of the parcels in each
point.
}
