% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PolyCentroid.R
\name{PolyCentroid}
\alias{PolyCentroid}
\title{PolyCentroid
Take a spatialpolygonsdataframe object and generate a spatialpointsdataframe
of their spatial centers. Preserve the attributes of the parcels in each point.}
\usage{
PolyCentroid(polygon)
}
\arguments{
\item{polygon}{A spatialpolygonsdataframe.}
}
\value{
"centroids" A spatialpointsdataframe of the centerspoints of the polygons.
}
\description{
PolyCentroid
Take a spatialpolygonsdataframe object and generate a spatialpointsdataframe
of their spatial centers. Preserve the attributes of the parcels in each point.
}
